2.8 Blueprint Classes & Instances
Don't ask where the other 7 commits went. Ran into problems and had to delete the repository.

Note to self: UE materials made in the engine have an apply button that must be pressed every time you change the material. This is different than other inspectors where the change is applied for you, and/or you safe it.