Section Wrapup - Simple Commit
This commit will finalize my work on Warehouse Wreckage. Normally, I'd give the game a shakeup to make it a true video game. However, I have no inspiration for this game, and I am still missing many design concepts to make a game that I would consider high-quality.

Because of this, I am going to consider this project done. It has done its job:
-I learned how to navigate the UE windows
-I learned how to craete Blueprint classes (Prefabs)
-I started to learn how blueprint coding works, though I am not thrilled by it.
-I learned how to translate C# programming into Blueprint coding.
-I learned how to set up a parent-child hierarchy for actors in a scene.
-And more

Hopefully, by the end of this course, I will have a game made that I feel like expanding upon. For this, I will consider this section a mini-project that taught the fundamentals of using UE5.